# web/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
      "AnnualOut": {
        "properties": {
          "tithi_pravesha": {
            "$ref": "#/components/schemas/TithiPraveshaOut"
          },
          "varshaphal": {
            "$ref": "#/components/schemas/VarshaphalOut"
          }
        },
        "required": [
          "varshaphal",
          "tithi_pravesha"
        ],
        "title": "AnnualOut",
        "type": "object"
      },
...
      "AnnualRequest": {
        "properties": {
          "birth": {
            "$ref": "#/components/schemas/BirthInput"
          },
          "place": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlaceInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "preset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Preset"
              },
              {
                "type": "null"
              }
            ]
          },
          "settings": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Settings"
              },
              {
                "type": "null"
              }
            ]
          },
          "years_completed": {
            "maximum": 120.0,
            "minimum": 0.0,
            "title": "Years Completed",
            "type": "integer"
          }
        },
        "required": [
          "birth",
          "years_completed"
        ],
        "title": "AnnualRequest",
        "type": "object"
      },
...
      "ChartRequest": {
        "properties": {
          "birth": {
            "$ref": "#/components/schemas/BirthInput"
          },
          "preset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Preset"
              },
              {
                "type": "null"
              }
            ]
          },
          "settings": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Settings"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "birth"
        ],
        "title": "ChartRequest",
        "type": "object"
      },
      "ChartResult": {
        "properties": {
          "ascendant": {
            "$ref": "#/components/schemas/PointOut"
          },
          "ayanamsa": {
            "$ref": "#/components/schemas/AyanamsaOut"
          },
          "birth": {
            "$ref": "#/components/schemas/BirthInput"
          },
          "dashas": {
            "$ref": "#/components/schemas/DashasOut"
          },
          "day": {
            "$ref": "#/components/schemas/DayOut"
          },
          "engine_version": {
            "title": "Engine Version",
            "type": "string"
          },
          "ephemeris": {
            "$ref": "#/components/schemas/EphemerisOut"
          },
          "grahas": {
            "items": {
              "$ref": "#/components/schemas/GrahaOut"
            },
            "title": "Grahas",
            "type": "array"
          },
          "houses": {
            "$ref": "#/components/schemas/HousesOut"
          },
          "midheaven": {
            "$ref": "#/components/schemas/PointOut"
          },
          "rodden_rating": {
            "title": "Rodden Rating",
            "type": "string"
          },
          "settings": {
            "$ref": "#/components/schemas/Settings"
          },
          "settings_hash": {
            "title": "Settings Hash",
            "type": "string"
          },
          "special": {
            "$ref": "#/components/schemas/SpecialPointsOut"
          },
          "time": {
            "$ref": "#/components/schemas/TimeOut"
          },
          "vargas": {
            "items": {
              "$ref": "#/components/schemas/VargaChartOut"
            },
            "title": "Vargas",
            "type": "array"
          }
        },
        "required": [
          "engine_version",
          "ephemeris",
          "settings",
          "settings_hash",
          "birth",
          "rodden_rating",
          "time",
          "ayanamsa",
          "ascendant",
          "midheaven",
          "grahas",
          "houses",
          "vargas",
          "special",
          "day",
          "dashas"
        ],
        "title": "ChartResult",
        "type": "object"
      },
...
      "DashaOut": {
        "properties": {
          "kalachakra_periods": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/KalachakraPeriod"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kalachakra Periods"
          },
          "sign_periods": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SignPeriod"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sign Periods"
          },
          "system": {
            "title": "System",
            "type": "string"
          },
          "table": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DashaTableOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "system"
        ],
        "title": "DashaOut",
        "type": "object"
      },
...
      "DashaRequest": {
        "properties": {
          "birth": {
            "$ref": "#/components/schemas/BirthInput"
          },
          "depth": {
            "default": 2,
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Depth",
            "type": "integer"
          },
          "preset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Preset"
              },
              {
                "type": "null"
              }
            ]
          },
          "settings": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Settings"
              },
              {
                "type": "null"
              }
            ]
          },
          "system": {
            "title": "System",
            "type": "string"
          }
        },
        "required": [
          "birth",
          "system"
        ],
        "title": "DashaRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KpChartOut": {
        "description": "KP chart: cusps with their lords, planets, significators, ruling planets.",
        "properties": {
          "cusps": {
            "items": {
              "$ref": "#/components/schemas/KpCuspOut"
            },
            "title": "Cusps",
            "type": "array"
          },
          "day_lord": {
            "$ref": "#/components/schemas/Body"
          },
          "horary_cusp_jd_ut": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Horary Cusp Jd Ut"
          },
          "horary_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Horary Number"
          },
          "house_significators": {
            "items": {
              "items": {
                "items": {
                  "$ref": "#/components/schemas/Body"
                },
                "type": "array"
              },
              "type": "array"
            },
            "title": "House Significators",
            "type": "array"
          },
          "planets": {
            "items": {
              "$ref": "#/components/schemas/KpPlanetOut"
            },
            "title": "Planets",
            "type": "array"
          },
          "ruling_planets": {
            "items": {
              "$ref": "#/components/schemas/Body"
            },
            "title": "Ruling Planets",
            "type": "array"
          }
        },
        "required": [
          "cusps",
          "planets",
          "house_significators",
          "day_lord",
          "ruling_planets"
        ],
        "title": "KpChartOut",
        "type": "object"
      },
...
      "Place": {
        "properties": {
          "admin1_code": {
            "title": "Admin1 Code",
            "type": "string"
          },
          "country_code": {
            "title": "Country Code",
            "type": "string"
          },
          "geonames_id": {
            "title": "Geonames Id",
            "type": "integer"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "population": {
            "title": "Population",
            "type": "integer"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "geonames_id",
          "name",
          "country_code",
          "admin1_code",
          "latitude",
          "longitude",
          "timezone",
          "population"
        ],
        "title": "Place",
        "type": "object"
      },
...
      "Preset": {
        "enum": [
          "classic_parashari",
          "drik_compatible",
          "kp",
          "pvr_jhora_style"
        ],
        "title": "Preset",
        "type": "string"
      },
...
      "SaturnTransit": {
        "enum": [
          "sade_sati",
          "ardhashtama",
          "ashtama",
          "kantaka"
        ],
        "title": "SaturnTransit",
        "type": "string"
      },
...
      "StrengthsOut": {
        "properties": {
          "ashtakavarga": {
            "$ref": "#/components/schemas/AshtakavargaOut"
          },
          "bhava_bala": {
            "items": {
              "$ref": "#/components/schemas/BhavaBalaOut"
            },
            "title": "Bhava Bala",
            "type": "array"
          },
          "shadbala": {
            "items": {
              "$ref": "#/components/schemas/ShadbalaOut"
            },
            "title": "Shadbala",
            "type": "array"
          },
          "vimshopaka": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "number"
              },
              "propertyNames": {
                "$ref": "#/components/schemas/Body"
              },
              "type": "object"
            },
            "title": "Vimshopaka",
            "type": "object"
          }
        },
        "required": [
          "shadbala",
          "bhava_bala",
          "ashtakavarga",
          "vimshopaka"
        ],
        "title": "StrengthsOut",
        "type": "object"
      },
...
      "TransitRequest": {
        "properties": {
          "birth": {
            "$ref": "#/components/schemas/BirthInput"
          },
          "end": {
            "format": "date",
            "title": "End",
            "type": "string"
          },
          "preset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Preset"
              },
              {
                "type": "null"
              }
            ]
          },
          "settings": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Settings"
              },
              {
                "type": "null"
              }
            ]
          },
          "start": {
            "format": "date",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "birth",
          "start",
          "end"
        ],
        "title": "TransitRequest",
        "type": "object"
      },
...
      "TransitsOut": {
        "properties": {
          "double_from_lagna": {
            "items": {
              "$ref": "#/components/schemas/DoubleTransit"
            },
            "title": "Double From Lagna",
            "type": "array"
          },
          "double_from_moon": {
            "items": {
              "$ref": "#/components/schemas/DoubleTransit"
            },
            "title": "Double From Moon",
            "type": "array"
          },
          "saturn": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/TransitEpisode"
              },
              "type": "array"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/SaturnTransit"
            },
            "title": "Saturn",
            "type": "object"
          }
        },
        "required": [
          "saturn",
          "double_from_lagna",
          "double_from_moon"
        ],
        "title": "TransitsOut",
        "type": "object"
      },
...
      "YogasOut": {
        "properties": {
          "cancelled": {
            "items": {
              "$ref": "#/components/schemas/YogaOut"
            },
            "title": "Cancelled",
            "type": "array"
          },
          "catalogue_size": {
            "title": "Catalogue Size",
            "type": "integer"
          },
          "present": {
            "items": {
              "$ref": "#/components/schemas/YogaOut"
            },
            "title": "Present",
            "type": "array"
          },
          "undecided": {
            "items": {
              "$ref": "#/components/schemas/UndecidedRuleOut"
            },
            "title": "Undecided",
            "type": "array"
          }
        },
        "required": [
          "catalogue_size",
          "present",
          "cancelled",
          "undecided"
        ],
        "title": "YogasOut",
        "type": "object"
      },
      "YogasRequest": {
        "properties": {
          "birth": {
            "$ref": "#/components/schemas/BirthInput"
          },
          "gender": {
            "anyOf": [
              {
                "enum": [
                  "male",
                  "female"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "include_sensitive": {
            "default": false,
            "title": "Include Sensitive",
            "type": "boolean"
          },
          "preset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Preset"
              },
              {
                "type": "null"
              }
            ]
          },
          "settings": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Settings"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "birth"
        ],
        "title": "YogasRequest",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Vedic astrology calculations with cited, explainable results.",
    "title": "Jyotish API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/charts": {
      "post": {
        "description": "The birth chart: positions, houses, divisional charts, special points, Vimshottari.",
        "operationId": "chart_v1_charts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chart"
      }
    },
    "/v1/charts/annual": {
      "post": {
        "description": "The Tajika annual chart and Tithi Pravesha for the year after ``years_completed``.",
        "operationId": "annual_v1_charts_annual_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnualRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnualOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Annual"
      }
    },
    "/v1/charts/dashas": {
      "post": {
        "description": "Any supported dasha system: a nakshatra dasha to ``depth`` levels, a sign dasha,\nor Kalachakra.",
        "operationId": "dashas_v1_charts_dashas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashaOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dashas"
      }
    },
    "/v1/charts/kp": {
      "post": {
        "description": "KP analysis; uses the KP preset unless settings or a preset are given.",
        "operationId": "kp_v1_charts_kp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KpChartOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Kp"
      }
    },
    "/v1/charts/strengths": {
      "post": {
        "operationId": "strengths_v1_charts_strengths_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrengthsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Strengths"
      }
    },
    "/v1/charts/transits": {
      "post": {
        "operationId": "transits_v1_charts_transits_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transits"
      }
    },
    "/v1/charts/yogas": {
      "post": {
        "operationId": "yogas_v1_charts_yogas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YogasRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YogasOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Yogas"
      }
    },
    "/v1/geo/search": {
      "get": {
        "description": "Places by name (GeoNames), most populous first.",
        "operationId": "geo_search_v1_geo_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 2,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2,
                  "minLength": 2,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Place"
                  },
                  "title": "Response Geo Search V1 Geo Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Geo Search"
      }
    },
    "/v1/kp/horary": {
      "post": {
        "operationId": "kp_horary_v1_kp_horary_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HoraryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KpChartOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Kp Horary"
      }
    },
    "/v1/match": {
      "post": {
        "operationId": "match_v1_match_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Match"
      }
    },
...
    "/v1/panchanga": {
      "post": {
        "operationId": "panchanga_v1_panchanga_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PanchangaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanchangaOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Panchanga"
      }
    },
...
  }
}